Feat: 예약하기 모달 세부정보 props 부모에서 내려주도록 함
