+++ FE/TAEK/src/components/main/rooms/reservation/ReservationModalItem.jsx
@@ -1,9 +1,10 @@
 import React from 'react'
+import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { numberComma, dayCounter } from 'utils/util';
+import { numberComma, dayCounter, formatDate } from 'utils/util';
 import ReservationFilterInfo from './ReservationFilterInfo';
 import ReservationChargeInfo from './ReservationChargeInfo';
-import ratingStar from 'public/images/rating-star.svg';
+import { MAIN } from 'constants/constant';
 
 const ReservationModalItemWrap = styled.div`
     padding: 30px 30px;
@@ -91,7 +92,20 @@ const CloseButton = styled.div`
 `;
 
 const ReservationModalItem = ({ handleModalToggle, roomData }) => {
+    const person = useSelector(({ person }) => person);
+    const { checkInDateInfo, checkOutDateInfo } = useSelector(({ date }) => date);
+    const { totalCount } = person;
     const { currentPrice, hotelRating } = roomData;
+
+    const checkInDateInfoText = `${formatDate(checkInDateInfo.year, checkInDateInfo.month, checkInDateInfo.day, '. ')}`;
+    const checkOutDateInfoText = `${formatDate(checkOutDateInfo.year, checkOutDateInfo.month, checkOutDateInfo.day, '. ')}`;
+
+    const dayCount = dayCounter(checkInDateInfo, checkOutDateInfo);
+    const totalRoomCharge = currentPrice * dayCount * totalCount;
+    const cleanUpCost = MAIN.RESERVATION.CLEAN_UP_COST;
+    const serviceCost = MAIN.RESERVATION.SERVICE_COST;
+    const tax = Math.floor(totalRoomCharge * MAIN.RESERVATION.TAX_RATE);
+    const totalCharge = totalRoomCharge + cleanUpCost + serviceCost + tax;
 
     return (
         <>
@@ -101,8 +115,8 @@ const ReservationModalItem = ({ handleModalToggle, roomData }) => {
                     <span className='currentPrice'>&#8361; {numberComma(currentPrice)}</span>
                     <span>/1박</span>
                 </div>
-                <ReservationFilterInfo {...{ ratingStar, hotelRating }} />
-                <ReservationChargeInfo {...{ currentPrice, numberComma, dayCounter }} />
+                <ReservationFilterInfo {...{ hotelRating, checkInDateInfoText, checkOutDateInfoText, person }} />
+                <ReservationChargeInfo {...{ numberComma, dayCount, totalCount, totalRoomCharge, cleanUpCost, serviceCost, tax, totalCharge }} />
                 <button className='reservation-btn'>예약하기</button>
                 <div className='reservation-info-text'>예약 확정 전에는 요금이 청구되지 않습니다.</div>
             </ReservationModalItemWrap>
